+++ src/app/test/requests_test.py
@@ -2,19 +2,35 @@
 from datetime import datetime, timedelta
 
 import pytest
+from copy import deepcopy
 from box import Box
 from brewtils.errors import ModelValidationError
 from brewtils.models import Choices, Command, Event, Events, Parameter
+from brewtils.models import Garden as BrewtilsGarden
 from brewtils.models import Request as BrewtilsRequest
 from mock import Mock, call, patch
 from mongomock.gridfs import enable_gridfs_integration
 
 import beer_garden.config
 import beer_garden.requests
-from beer_garden.db.mongo.models import Request
+from beer_garden.db.mongo.models import Request, Garden
 from beer_garden.requests import RequestValidator
+from beer_garden.garden import create_garden, get_garden
 
 enable_gridfs_integration()
+
+
+@pytest.fixture(autouse=True)
+def drop():
+    yield
+    Garden.drop_collection()
+
+
+@pytest.fixture
+def localgarden():
+    yield create_garden(
+        BrewtilsGarden(name="parent", connection_type="LOCAL", systems=[])
+    )
 
 
 @pytest.fixture
@@ -44,6 +60,22 @@ def validator(monkeypatch, mongo_conn):
 
 @pytest.fixture
 def child_garden_request():
+    request = Request(
+        namespace="child_garden",
+        system="testsystem",
+        system_version="1.0.0",
+        instance_name="instance1",
+        command="somecommand",
+        parameters={},
+    )
+    request.save()
+
+    yield request
+    request.delete()
+
+
+@pytest.fixture
+def latency_request():
     request = Request(
         namespace="child_garden",
         system="testsystem",
@@ -1092,11 +1124,126 @@ class TestHandleEvent:
         child_garden_request.status = "SUCCESS"
         child_garden_request.status_updated_at = status_updated_at
         request_event = Event(
-            payload=child_garden_request, name=Events.REQUEST_UPDATED.name
+            payload=child_garden_request,
+            name=Events.REQUEST_UPDATED.name,
+            garden="child",
         )
+
+        beer_garden.config._CONFIG = {
+            "metrics": {"garden_latency_metrics_cache_window": 60}
+        }
 
         beer_garden.requests.handle_event(request_event)
 
         updated_request = Request.objects.get(id=child_garden_request.id)
 
         assert updated_request.status_updated_at == status_updated_at
+
+
+class TestLatencyRequest:
+    def test_create_request(self, latency_request):
+        latency_request.status = "CREATE"
+        latency_request.source_garden = "parent"
+        latency_request.target_garden = "target"
+
+        updated_request = deepcopy(latency_request)
+        status_updated_at = datetime.utcnow() + timedelta(minutes=1)
+        status_updated_at = status_updated_at.replace(microsecond=0)
+        updated_request.created_at = status_updated_at
+
+        request_event = Event(payload=updated_request, name=Events.REQUEST_CREATED.name)
+
+        beer_garden.requests.update_request_latency_garden(
+            latency_request, request_event
+        )
+
+        assert beer_garden.requests.cached_latency_metrics
+
+        assert "parent" in beer_garden.requests.cached_latency_metrics
+        assert "target" in beer_garden.requests.cached_latency_metrics["parent"]
+        assert (
+            len(
+                beer_garden.requests.cached_latency_metrics["parent"]["target"][
+                    "CREATE"
+                ]
+            )
+            > 0
+        )
+
+    def test_in_progress_request(self, latency_request):
+        latency_request.status = "IN_PROGRESS"
+        latency_request.source_garden = "parent"
+        latency_request.target_garden = "target"
+
+        updated_request = deepcopy(latency_request)
+        status_updated_at = datetime.utcnow() + timedelta(minutes=1)
+        status_updated_at = status_updated_at.replace(microsecond=0)
+        updated_request.created_at = status_updated_at
+
+        request_event = Event(payload=updated_request, name=Events.REQUEST_UPDATED.name)
+
+        beer_garden.requests.update_request_latency_garden(
+            latency_request, request_event
+        )
+
+        assert beer_garden.requests.cached_latency_metrics
+
+        assert "parent" in beer_garden.requests.cached_latency_metrics
+        assert "target" in beer_garden.requests.cached_latency_metrics["parent"]
+        assert (
+            len(
+                beer_garden.requests.cached_latency_metrics["parent"]["target"]["START"]
+            )
+            > 0
+        )
+
+    def test_complete_request(self, latency_request):
+        latency_request.status = "SUCCESS"
+        latency_request.source_garden = "parent"
+        latency_request.target_garden = "target"
+
+        updated_request = deepcopy(latency_request)
+        status_updated_at = datetime.utcnow() + timedelta(minutes=1)
+        status_updated_at = status_updated_at.replace(microsecond=0)
+        updated_request.created_at = status_updated_at
+
+        request_event = Event(payload=updated_request, name=Events.REQUEST_UPDATED.name)
+
+        beer_garden.requests.update_request_latency_garden(
+            latency_request, request_event
+        )
+
+        assert beer_garden.requests.cached_latency_metrics
+
+        assert "parent" in beer_garden.requests.cached_latency_metrics
+        assert "target" in beer_garden.requests.cached_latency_metrics["parent"]
+        assert (
+            len(
+                beer_garden.requests.cached_latency_metrics["parent"]["target"][
+                    "COMPLETE"
+                ]
+            )
+            > 0
+        )
+
+    def test_update_cache(self, localgarden, latency_request):
+        latency_request.status = "CREATE"
+        latency_request.source_garden = "parent"
+        latency_request.target_garden = "target"
+
+        updated_request = deepcopy(latency_request)
+        status_updated_at = datetime.utcnow() + timedelta(minutes=1)
+        status_updated_at = status_updated_at.replace(microsecond=0)
+        updated_request.created_at = status_updated_at
+
+        request_event = Event(payload=updated_request, name=Events.REQUEST_CREATED.name)
+
+        beer_garden.requests.update_request_latency_garden(
+            latency_request, request_event
+        )
+
+        beer_garden.requests.store_cache_latency_metrics()
+
+        updated_garden = get_garden(localgarden.name)
+
+        assert updated_garden.metadata["CREATE_COUNT_target"] > 0