+++ CHANGELOG.md
@@ -5,6 +5,7 @@
 TBD
 
 - Fixed bug where removing a parent garden failed to remove both parent and child
+- Added caching to Request Latency Metrics for Gardens to reduce database calls
 - Added flag to enable API heartbeat checks
 - Added Topics API and classes to support dynamic topic subscribers
 - Now show alert on Garden Admin page when a Sync event is seen